Revert "UBUNTU: SAUCE: fix pv-ops for legacy Xen"
This reverts commit 3f3fba59aa5773836d94799d10b692f9b7ea16a0.

It turned out that since 2.6.39 the Xen code in the kernel would
not try to set OSXSAVE in CR4 to find out whether the feature is
supported. Instead it relies on OSXSAVE being set by the HV in
the cpuid bits (which will only happen if supported).

BugLink: http://bugs.launchpad.net/bugs/1044550

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Colin King <colin.king@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>