stable/sru-workflow-manager: introduce kernel-stable-master-bug property
Add kernel-stable-master-bug property to derivative tracking bugs. The
property is set to the tracking bug number of the master branch of that
derivative. It'll be used in the following change to track which bug
this derivative is following, so we can prevent setting the promote-*
tasks before the master is ready.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>