stable/sru-workflow-manager: don't confirm promote-* tasks before master
It was agreed that we shouldn't set promote-to-* tasks to Confirmed on
derivative tracking bugs when the tracking bug for the master branch
isn't promoted yet. This adds a function which checks the relevant task
status on the master tracking bug, and use it where needed to only set
promote-to-* tasks when master has the same task as Fix Released.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>