Add master bug information also to tracking bugs for backports
Tracking bugs for linux backport packages also should wait on the
master linux promotion to proposed or updates, and should have master
tracking bug information, like previous changes for derivative bugs.
Move the setting of kernel-stable-master-bug to a common place (the
opening code in TrackingBug class), and pass the master bug number to
it also when opening the tracking bug for any backport package.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>