stable/create-release-tracker: add --master_bug option
If creating a tracking bug for a derivative or backport branch manually,
allow passing the master bug information through command line with this
new option. Having the master bug information on the tracking bug is the
only way shank bot has to know precisely from which master release the
bug derives so it can properly wait on promotion tasks when needed.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>